Refactor MessageChunk into a separate module.
Moved the `MessageChunk` class from `chat.py` to a new file `chat_batch.py` to improve modularity and maintainability. This change separates responsibilities, making the codebase more organized and easier to extend.
